requirements.txt
pynput==1.7.6
python-vlc==3.0.18121
wakepy==0.7.2